Removal of legacy code and major bug fix
I just realized that if you put capital letters in your channel name it would crash the entire game. This is part of a bigger issue but i have implemented fix that should work for all of you but i would definitely recommend having auto-save on in the time being. If you are still crashing please submit output_logs to me.